Add stripped down version of dumbell controller and tests
